+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/room/send/LocalEchoEventFactory.kt
@@ -23,6 +23,7 @@ import androidx.exifinterface.media.ExifInterface
 import im.vector.matrix.android.R
 import im.vector.matrix.android.api.permalinks.PermalinkFactory
 import im.vector.matrix.android.api.session.content.ContentAttachmentData
+import im.vector.matrix.android.api.session.events.model.Content
 import im.vector.matrix.android.api.session.events.model.Event
 import im.vector.matrix.android.api.session.events.model.EventType
 import im.vector.matrix.android.api.session.events.model.LocalEcho
@@ -95,7 +96,7 @@ internal class LocalEchoEventFactory @Inject constructor(
             return createFormattedTextEvent(roomId, createTextContent(text, autoMarkdown), msgType)
         }
         val content = MessageTextContent(msgType = msgType, body = text.toString())
-        return createEvent(roomId, content)
+        return createMessageEvent(roomId, content)
     }
 
     private fun createTextContent(text: CharSequence, autoMarkdown: Boolean): TextContent {
@@ -129,7 +130,7 @@ internal class LocalEchoEventFactory @Inject constructor(
             text != htmlText && htmlText != "<p>${text.trim()}</p>\n"
 
     fun createFormattedTextEvent(roomId: String, textContent: TextContent, msgType: String): Event {
-        return createEvent(roomId, textContent.toMessageTextContent(msgType))
+        return createMessageEvent(roomId, textContent.toMessageTextContent(msgType))
     }
 
     fun createReplaceTextEvent(roomId: String,
@@ -138,7 +139,7 @@ internal class LocalEchoEventFactory @Inject constructor(
                                newBodyAutoMarkdown: Boolean,
                                msgType: String,
                                compatibilityText: String): Event {
-        return createEvent(roomId,
+        return createMessageEvent(roomId,
                 MessageTextContent(
                         msgType = msgType,
                         body = compatibilityText,
@@ -153,7 +154,7 @@ internal class LocalEchoEventFactory @Inject constructor(
                                 pollEventId: String,
                                 optionIndex: Int,
                                 optionLabel: String): Event {
-        return createEvent(roomId,
+        return createMessageEvent(roomId,
                 MessagePollResponseContent(
                         body = optionLabel,
                         relatesTo = RelationDefaultContent(
@@ -175,7 +176,7 @@ internal class LocalEchoEventFactory @Inject constructor(
                 append(it.value)
             }
         }
-        return createEvent(
+        return createMessageEvent(
                 roomId,
                 MessageOptionsContent(
                         body = compatLabel,
@@ -211,7 +212,7 @@ internal class LocalEchoEventFactory @Inject constructor(
         //
         val replyFallback = buildReplyFallback(body, originalEvent.root.senderId ?: "", newBodyText)
 
-        return createEvent(roomId,
+        return createMessageEvent(roomId,
                 MessageTextContent(
                         msgType = msgType,
                         body = compatibilityText,
@@ -280,7 +281,7 @@ internal class LocalEchoEventFactory @Inject constructor(
                 ),
                 url = attachment.queryUri.toString()
         )
-        return createEvent(roomId, content)
+        return createMessageEvent(roomId, content)
     }
 
     private fun createVideoEvent(roomId: String, attachment: ContentAttachmentData): Event {
@@ -316,7 +317,7 @@ internal class LocalEchoEventFactory @Inject constructor(
                 ),
                 url = attachment.queryUri.toString()
         )
-        return createEvent(roomId, content)
+        return createMessageEvent(roomId, content)
     }
 
     private fun createAudioEvent(roomId: String, attachment: ContentAttachmentData): Event {
@@ -329,7 +330,7 @@ internal class LocalEchoEventFactory @Inject constructor(
                 ),
                 url = attachment.queryUri.toString()
         )
-        return createEvent(roomId, content)
+        return createMessageEvent(roomId, content)
     }
 
     private fun createFileEvent(roomId: String, attachment: ContentAttachmentData): Event {
@@ -342,18 +343,22 @@ internal class LocalEchoEventFactory @Inject constructor(
                 ),
                 url = attachment.queryUri.toString()
         )
-        return createEvent(roomId, content)
+        return createMessageEvent(roomId, content)
     }
 
-    private fun createEvent(roomId: String, content: Any? = null): Event {
+    private fun createMessageEvent(roomId: String, content: MessageContent? = null): Event {
+        return createEvent(roomId, EventType.MESSAGE, content.toContent())
+    }
+
+    fun createEvent(roomId: String, type: String, content: Content?): Event {
         val localId = LocalEcho.createLocalEchoId()
         return Event(
                 roomId = roomId,
                 originServerTs = dummyOriginServerTs(),
                 senderId = userId,
                 eventId = localId,
-                type = EventType.MESSAGE,
-                content = content.toContent(),
+                type = type,
+                content = content,
                 unsignedData = UnsignedData(age = null, transactionId = localId)
         )
     }
@@ -410,7 +415,7 @@ internal class LocalEchoEventFactory @Inject constructor(
                 formattedBody = replyFormatted,
                 relatesTo = RelationDefaultContent(null, null, ReplyToContent(eventId))
         )
-        return createEvent(roomId, content)
+        return createMessageEvent(roomId, content)
     }
 
     private fun buildReplyFallback(body: TextContent, originalSenderId: String?, newBodyText: String): String {