+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailFragment.kt
@@ -26,6 +26,7 @@ import android.os.Bundle
 import android.os.Parcelable
 import android.text.Spannable
 import android.view.HapticFeedbackConstants
+import android.view.LayoutInflater
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
@@ -65,6 +66,7 @@ import im.vector.matrix.android.api.permalinks.PermalinkFactory
 import im.vector.matrix.android.api.session.Session
 import im.vector.matrix.android.api.session.content.ContentAttachmentData
 import im.vector.matrix.android.api.session.events.model.Event
+import im.vector.matrix.android.api.session.events.model.toModel
 import im.vector.matrix.android.api.session.file.FileService
 import im.vector.matrix.android.api.session.room.model.Membership
 import im.vector.matrix.android.api.session.room.model.message.MessageAudioContent
@@ -72,6 +74,7 @@ import im.vector.matrix.android.api.session.room.model.message.MessageContent
 import im.vector.matrix.android.api.session.room.model.message.MessageFileContent
 import im.vector.matrix.android.api.session.room.model.message.MessageFormat
 import im.vector.matrix.android.api.session.room.model.message.MessageImageInfoContent
+import im.vector.matrix.android.api.session.room.model.message.MessageStickerContent
 import im.vector.matrix.android.api.session.room.model.message.MessageTextContent
 import im.vector.matrix.android.api.session.room.model.message.MessageVerificationRequestContent
 import im.vector.matrix.android.api.session.room.model.message.MessageVideoContent
@@ -80,6 +83,7 @@ import im.vector.matrix.android.api.session.room.send.SendState
 import im.vector.matrix.android.api.session.room.timeline.Timeline
 import im.vector.matrix.android.api.session.room.timeline.TimelineEvent
 import im.vector.matrix.android.api.session.room.timeline.getLastMessageContent
+import im.vector.matrix.android.api.session.widgets.model.WidgetType
 import im.vector.matrix.android.api.util.MatrixItem
 import im.vector.matrix.android.api.util.toMatrixItem
 import im.vector.matrix.android.internal.crypto.attachments.toElementToDecrypt
@@ -131,6 +135,7 @@ import im.vector.riotx.features.crypto.verification.VerificationBottomSheet
 import im.vector.riotx.features.home.AvatarRenderer
 import im.vector.riotx.features.home.room.detail.composer.TextComposerView
 import im.vector.riotx.features.home.room.detail.readreceipts.DisplayReadReceiptsBottomSheet
+import im.vector.riotx.features.home.room.detail.sticker.StickerPickerConstants
 import im.vector.riotx.features.home.room.detail.timeline.TimelineEventController
 import im.vector.riotx.features.home.room.detail.timeline.action.EventSharedAction
 import im.vector.riotx.features.home.room.detail.timeline.action.MessageActionsBottomSheet
@@ -143,6 +148,8 @@ import im.vector.riotx.features.home.room.detail.timeline.item.MessageInformatio
 import im.vector.riotx.features.home.room.detail.timeline.item.MessageTextItem
 import im.vector.riotx.features.home.room.detail.timeline.item.ReadReceiptData
 import im.vector.riotx.features.home.room.detail.timeline.reactions.ViewReactionsBottomSheet
+import im.vector.riotx.features.home.room.detail.widget.RoomWidgetsBannerView
+import im.vector.riotx.features.home.room.detail.widget.RoomWidgetsBottomSheet
 import im.vector.riotx.features.html.EventHtmlRenderer
 import im.vector.riotx.features.html.PillImageSpan
 import im.vector.riotx.features.invite.VectorInviteView
@@ -155,6 +162,7 @@ import im.vector.riotx.features.reactions.EmojiReactionPickerActivity
 import im.vector.riotx.features.settings.VectorPreferences
 import im.vector.riotx.features.share.SharedData
 import im.vector.riotx.features.themes.ThemeUtils
+import im.vector.riotx.features.widgets.WidgetActivity
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 import kotlinx.android.parcel.Parcelize
@@ -194,7 +202,8 @@ class RoomDetailFragment @Inject constructor(
         VectorInviteView.Callback,
         JumpToReadMarkerView.Callback,
         AttachmentTypeSelectorView.Callback,
-        AttachmentsHelper.Callback {
+        AttachmentsHelper.Callback,
+        RoomWidgetsBannerView.Callback {
 
     companion object {
 
@@ -259,6 +268,8 @@ class RoomDetailFragment @Inject constructor(
         setupNotificationView()
         setupJumpToReadMarkerView()
         setupJumpToBottomView()
+        setupWidgetsBannerView()
+
         roomToolbarContentView.debouncedClicks {
             navigator.openRoomProfile(requireActivity(), roomDetailArgs.roomId)
         }
@@ -289,18 +300,47 @@ class RoomDetailFragment @Inject constructor(
 
         roomDetailViewModel.observeViewEvents {
             when (it) {
-                is RoomDetailViewEvents.Failure                -> showErrorInSnackbar(it.throwable)
-                is RoomDetailViewEvents.OnNewTimelineEvents    -> scrollOnNewMessageCallback.addNewTimelineEventIds(it.eventIds)
-                is RoomDetailViewEvents.ActionSuccess          -> displayRoomDetailActionSuccess(it)
-                is RoomDetailViewEvents.ActionFailure          -> displayRoomDetailActionFailure(it)
-                is RoomDetailViewEvents.ShowMessage            -> showSnackWithMessage(it.message, Snackbar.LENGTH_LONG)
-                is RoomDetailViewEvents.NavigateToEvent        -> navigateToEvent(it)
-                is RoomDetailViewEvents.FileTooBigError        -> displayFileTooBigError(it)
-                is RoomDetailViewEvents.DownloadFileState      -> handleDownloadFileState(it)
-                is RoomDetailViewEvents.JoinRoomCommandSuccess -> handleJoinedToAnotherRoom(it)
-                is RoomDetailViewEvents.SendMessageResult      -> renderSendMessageResult(it)
+                is RoomDetailViewEvents.Failure                         -> showErrorInSnackbar(it.throwable)
+                is RoomDetailViewEvents.OnNewTimelineEvents             -> scrollOnNewMessageCallback.addNewTimelineEventIds(it.eventIds)
+                is RoomDetailViewEvents.ActionSuccess                   -> displayRoomDetailActionSuccess(it)
+                is RoomDetailViewEvents.ActionFailure                   -> displayRoomDetailActionFailure(it)
+                is RoomDetailViewEvents.ShowMessage                     -> showSnackWithMessage(it.message, Snackbar.LENGTH_LONG)
+                is RoomDetailViewEvents.NavigateToEvent                 -> navigateToEvent(it)
+                is RoomDetailViewEvents.FileTooBigError                 -> displayFileTooBigError(it)
+                is RoomDetailViewEvents.DownloadFileState               -> handleDownloadFileState(it)
+                is RoomDetailViewEvents.JoinRoomCommandSuccess          -> handleJoinedToAnotherRoom(it)
+                is RoomDetailViewEvents.SendMessageResult               -> renderSendMessageResult(it)
+                RoomDetailViewEvents.DisplayPromptForIntegrationManager -> displayPromptForIntegrationManager()
+                is RoomDetailViewEvents.OpenStickerPicker               -> openStickerPicker(it)
             }.exhaustive
         }
+    }
+
+    private fun setupWidgetsBannerView() {
+        roomWidgetsBannerView.callback = this
+    }
+
+    private fun openStickerPicker(event: RoomDetailViewEvents.OpenStickerPicker) {
+        navigator.openStickerPicker(this, roomDetailArgs.roomId, event.widget)
+    }
+
+    private fun displayPromptForIntegrationManager() {
+        // The Sticker picker widget is not installed yet. Propose the user to install it
+        val builder = AlertDialog.Builder(requireContext())
+        val v: View = LayoutInflater.from(requireContext()).inflate(R.layout.dialog_no_sticker_pack, null)
+        builder
+                .setView(v)
+                .setPositiveButton(R.string.yes) { _, _ ->
+                    // Open integration manager, to the sticker installation page
+                    navigator.openIntegrationManager(
+                            context = requireContext(),
+                            roomId = roomDetailArgs.roomId,
+                            integId = null,
+                            screen = WidgetType.StickerPicker.preferred
+                    )
+                }
+                .setNegativeButton(R.string.no, null)
+                .show()
     }
 
     private fun handleJoinedToAnotherRoom(action: RoomDetailViewEvents.JoinRoomCommandSuccess) {
@@ -428,18 +468,24 @@ class RoomDetailFragment @Inject constructor(
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        if (item.itemId == R.id.clear_message_queue) {
-            // This a temporary option during dev as it is not super stable
-            // Cancel all pending actions in room queue and post a dummy
-            // Then mark all sending events as undelivered
-            roomDetailViewModel.handle(RoomDetailAction.ClearSendQueue)
-            return true
+        return when (item.itemId) {
+            R.id.clear_message_queue -> {
+                // This a temporary option during dev as it is not super stable
+                // Cancel all pending actions in room queue and post a dummy
+                // Then mark all sending events as undelivered
+                roomDetailViewModel.handle(RoomDetailAction.ClearSendQueue)
+                true
+            }
+            R.id.resend_all          -> {
+                roomDetailViewModel.handle(RoomDetailAction.ResendAll)
+                true
+            }
+            R.id.open_matrix_apps    -> {
+                navigator.openIntegrationManager(requireContext(), roomDetailArgs.roomId, null, null)
+                true
+            }
+            else                     -> super.onOptionsItemSelected(item)
         }
-        if (item.itemId == R.id.resend_all) {
-            roomDetailViewModel.handle(RoomDetailAction.ResendAll)
-            return true
-        }
-        return super.onOptionsItemSelected(item)
     }
 
     private fun renderRegularMode(text: String) {
@@ -514,14 +560,18 @@ class RoomDetailFragment @Inject constructor(
         val hasBeenHandled = attachmentsHelper.onActivityResult(requestCode, resultCode, data)
         if (!hasBeenHandled && resultCode == RESULT_OK && data != null) {
             when (requestCode) {
-                AttachmentsPreviewActivity.REQUEST_CODE -> {
+                AttachmentsPreviewActivity.REQUEST_CODE            -> {
                     val sendData = AttachmentsPreviewActivity.getOutput(data)
                     val keepOriginalSize = AttachmentsPreviewActivity.getKeepOriginalSize(data)
                     roomDetailViewModel.handle(RoomDetailAction.SendMedia(sendData, !keepOriginalSize))
                 }
-                REACTION_SELECT_REQUEST_CODE            -> {
+                REACTION_SELECT_REQUEST_CODE                       -> {
                     val (eventId, reaction) = EmojiReactionPickerActivity.getOutput(data) ?: return
                     roomDetailViewModel.handle(RoomDetailAction.SendReaction(eventId, reaction))
+                }
+                StickerPickerConstants.STICKER_PICKER_REQUEST_CODE -> {
+                    val content = WidgetActivity.getOutput(data).toModel<MessageStickerContent>() ?: return
+                    roomDetailViewModel.handle(RoomDetailAction.SendSticker(content))
                 }
             }
         }
@@ -676,6 +726,7 @@ class RoomDetailFragment @Inject constructor(
         val summary = state.asyncRoomSummary()
         val inviter = state.asyncInviter()
         if (summary?.membership == Membership.JOIN) {
+            roomWidgetsBannerView.render(state.activeRoomWidgets())
             scrollOnHighlightedEventCallback.timeline = roomDetailViewModel.timeline
             timelineEventController.update(state)
             inviteView.visibility = View.GONE
@@ -1379,7 +1430,7 @@ class RoomDetailFragment @Inject constructor(
             AttachmentTypeSelectorView.Type.GALLERY -> attachmentsHelper.selectGallery(this)
             AttachmentTypeSelectorView.Type.AUDIO   -> attachmentsHelper.selectAudio(this)
             AttachmentTypeSelectorView.Type.CONTACT -> attachmentsHelper.selectContact(this)
-            AttachmentTypeSelectorView.Type.STICKER -> vectorBaseActivity.notImplemented("Adding stickers")
+            AttachmentTypeSelectorView.Type.STICKER -> roomDetailViewModel.handle(RoomDetailAction.SelectStickerAttachment)
         }.exhaustive
     }
 
@@ -1411,5 +1462,10 @@ class RoomDetailFragment @Inject constructor(
         super.onContactAttachmentReady(contactAttachment)
         val formattedContact = contactAttachment.toHumanReadable()
         roomDetailViewModel.handle(RoomDetailAction.SendMessage(formattedContact, false))
+    }
+
+    override fun onViewWidgetsClicked() {
+        RoomWidgetsBottomSheet.newInstance()
+                .show(childFragmentManager, "ROOM_WIDGETS_BOTTOM_SHEET")
     }
 }